improved IO for match_genes per usage on hodor
Adding the ks_convergence_intrawalk script.  This runs the kolmogorov
smirnov test on a set of walks.  It can be run after any number of steps
have been generated.  If more steps are added it can read the existing
calculations from a file to avoid redoing all of the calculations each
time a new file is added.